Move new defs to .interface
